requirements.txt
# Plant Management System - Python Dependencies

# ETL Pipeline
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
supabase>=2.0.0

# Backend (FastAPI)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Monitoring & Logging
structlog>=24.1.0
